*** empty log message ***
git-svn-id: svn://svn.savannah.nongnu.org/swarm/trunk@5623 84d386b0-d280-4341-8980-c4cb7e1eb7a5

+++ swarm/src/defobj/JavaCollectionIndex.h
@@ -3,11 +3,11 @@
 
 @interface JavaCollectionIndex: CreateDrop
 {
-  size_t count;
+  unsigned count;
   int pos;
 }
 #ifdef HAVE_JDK
-+ create: aZone setCount: (size_t)count;
++ create: aZone setCount: (unsigned)count;
 - (id <Symbol>)getLoc;
 - next;
 #endif
